# Reduce polymorphism resulting from unstable Node shapes