--- a/.gitignore
+++ b/.gitignore
@@ -84,3 +84,4 @@ tests/cases/user/puppeteer/puppeteer
 tests/cases/user/axios-src/axios-src
 tests/cases/user/prettier/prettier
 .eslintcache
+*v8.log