--- a/src/compiler/emitter.ts
+++ b/src/compiler/emitter.ts
@@ -2475,8 +2475,7 @@ export function createPrinter(printerOptions: PrinterOptions = {}, handlers: Pri
     //
 
     function emitPrivateIdentifier(node: PrivateIdentifier) {
-        const writeText = node.symbol ? writeSymbol : write;
-        writeText(getTextOfNode(node, /*includeTrivia*/ false), node.symbol);
+        write(getTextOfNode(node, /*includeTrivia*/ false));
     }