… src/compiler/checker.ts
… src/compiler/factory/nodeFactory.ts
… src/compiler/program.ts
… src/compiler/types.ts
… src/compiler/utilities.ts
… src/compiler/utilitiesPublic.ts
… src/harness/harnessUtils.ts
… src/services/codefixes/convertToAsyncFunction.ts
… src/services/codefixes/fixImplicitThis.ts
… src/services/codefixes/fixImportNonExportedMember.ts
… src/services/completions.ts
… src/services/documentHighlights.ts
… src/services/findAllReferences.ts
… src/services/goToDefinition.ts
… src/services/importTracker.ts
… src/services/inlayHints.ts
… src/services/refactors/convertExport.ts
… src/services/refactors/extractSymbol.ts
… src/services/refactors/moveToNewFile.ts
… src/services/signatureHelp.ts
… src/services/suggestionDiagnostics.ts
… src/services/symbolDisplay.ts
… src/services/utilities.ts
… tests/baselines/reference/api/tsserverlibrary.d.ts
… tests/baselines/reference/api/typescript.d.ts
… tests/baselines/reference/computedPropertyNames21_ES5.symbols
… tests/baselines/reference/computedPropertyNames21_ES5.types
… tests/baselines/reference/computedPropertyNames21_ES6.symbols
… tests/baselines/reference/computedPropertyNames21_ES6.types
… tests/baselines/reference/computedPropertyNames23_ES5.symbols