@rbuckton 2022-11-29
@typescript-bot perf test