@rbuckton 2022-12-06
@typescript-bot perf test

Running again after rebasing against `main` in case something changed.