@rbuckton 2022-12-07
@typescript-bot perf test