@weswigham 2022-12-07
Q: Is it possible for us to help enforce monomorphic object shapes using a lint rule of some kind? We basically want to enforce that we don't potentially add properties outside of the constructor function for an object, right? (And, as the symbol tracker shows, prefer explicit wrappers over object spread and dynamic functions)