@rbuckton 2022-12-08
> Q: Is it possible for us to help enforce monomorphic object shapes using a lint rule of some kind? We basically want to enforce that we don't potentially add properties outside of the constructor function for an object, right? (And, as the symbol tracker shows, prefer explicit wrappers over object spread and dynamic functions)

I was talking with Ryan about that today. I may be able to repurpose my V8 log processor to run each of our benchmark suites to collect map details per constructor and report on outsized map counts or maps created in files other than their expected origin.