@Andarist 2022-12-08
@Jack-Works I'm not sure how closely this relates to the stats posted by the bot here (most of them overlap but I'm not sure if they are 1 to 1) but you can call `tsc --diagnostics` and `tsc --extendedDiagnostics`