@rbuckton 2022-12-08
> I'm curious. How I can run the perf on my own project?

The tool we use for benchmarking isn't public, but it essentially just runs `tsc --extendedDiagnostics` repeatedly on various projects and summarizes the results.

I'm planning to publish the tool I wrote to analyze V8 maps and deoptimizations in the near future, however.