@Jack-Works 2022-12-09
System info:

Node 19.2.0, Windows 10.0.22623, AMD64 Family 25 Model 33 Stepping 0 AuthenticAMD ~3701 Mhz

Project scale:

```
Projects in scope:                         75
Projects built:                            71
Aggregate Files:                        64824
Aggregate Lines of Library:            716308
Aggregate Lines of Definitions:       8345732
Aggregate Lines of TypeScript:         216245
Aggregate Lines of JavaScript:              0
Aggregate Lines of JSON:                40878
Aggregate Lines of Other:                   0
```

Perf result:

5.0.0-dev.20221207 => npm:@typescript-deploys/pr-build@5.0.0-pr-51682-13

```
Aggregate Identifiers:                9923117
Aggregate Symbols:                    8689294 => 8689301
Aggregate Types:                       998455 => 998526
Aggregate Instantiations:             6990544
Aggregate Memory used:               2633329K => 2685920K
Aggregate Assignability cache size:    428849
Aggregate Identity cache size:          53133
Aggregate Subtype cache size:           49022
Aggregate Strict subtype cache size:    37922
Aggregate I/O Read time:                0.72s
Aggregate Parse time:                   2.23s => 2.32s
Aggregate ResolveModule time:           2.79s => 2.78s
Aggregate ResolveTypeReference time:    0.10s => 0.11s
Aggregate Program time:                10.15s => 10.23s
Aggregate Bind time:                    1.08s => 10.7s
Aggregate Check time:                  35.31s => 35.43s
Aggregate transformTime time:           5.92s => 5.81s
Aggregate Source Map time:              0.83s => 0.81s
Aggregate commentTime time:             1.19s => 0.16s
Aggregate printTime time:              19.76s => 19.60s
Aggregate Emit time:                   19.82s => 19.68s
Aggregate I/O Write time:               2.81s => 2.90s
Config file parsing time:               0.32s
Up-to-date check time:                  0.01s
Build time:                            90.30s => 90.60s
```